#36 Move telemetry to Concierge Core
